Commit all changes made in packadges model, controller,
also change in pom.xml related to HATEOS Additional dependency

10.04.2021

+++ lirabry/src/main/java/com/library/lirabry/servicesImpl/AuthorService.java
@@ -67,8 +67,13 @@ public class AuthorService implements AuthorServiceI {
 	}
 	
 	@Override
-	public void deleteAuthor(Long id) {
-		authorRepository.deleteById(id);
+	public boolean deleteAuthor(Long id) {
+		if(authorRepository.findById(id).isPresent()) {
+			authorRepository.deleteById(id);
+			return true;
+		}else {
+			return false;
+		}
 	}
 
 }
